+++ site/index.js
@@ -1,23 +1,8 @@
-const i18n = require("./i18n");
-
-// Process redirects and localization before running anything else
-(() => {
-    if (location.href.includes("heissepreise.github.io")) {
-        location.href = "https://heisse-preise.io";
-        return;
-    }
-
-    // Find the most preferred supported language
-    for (const langCode of navigator.languages) {
-        // We don't do regional codes, so take just the language code
-        let lang = langCode.length >= 2 ? langCode.substring(0, 2) : null;
-        if (lang == null) continue;
-        if (i18n.locales.includes(lang)) {
-            i18n.setLocale(lang);
-            break;
-        }
-    }
-})();
+// Process redirects before running anything else
+if (location.href.includes("heissepreise.github.io")) {
+    location.href = "https://heisse-preise.io";
+    return;
+}
 
 const { getQueryParameter } = require("./js/misc");
 const model = require("./model");
